requirements.txt
streamlit==1.45.0
pandas==2.2.3
pydantic==2.10.6
pytest==8.3.4
fastapi==0.103.1
mysql-connector-python==9.4.0
requests==2.31.0
uvicorn==0.23.2